requirements.txt
Flask==3.1.2
python-dotenv==1.2.1
langchain-core==0.3.75
langchain-groq==0.3.7
pdfplumber==0.11.7